refactor: improve not found interface
