+++ src/features/module/pages/form-introduction-page.tsx
@@ -1,7 +1,7 @@
 import { ROUTES } from '@/app/configs/routes';
 import { Editor } from '@/components/editor';
-import { NotFound } from '@/components/not-found';
 import { Button } from '@/components/ui/button';
+import { NotFound } from '@/templates/not-found';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useEffect } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
